tools: docker: test_flows: use mktemp instead of tempfile
tempfile is deprecated and doesn't exist on non-Debian systems. Replace
it with mktemp.

While we're at it, check if mktemp failed and exit.

+++ tools/docker/tests/test_flows.sh
@@ -158,7 +158,8 @@ test_init() {
         exit 1
     }
     #save bml_conn_map output into a file.
-    connmap=$(tempfile)
+    connmap=$(mktemp)
+    [ -z "$connmap" ] && { err "Failed to create temp file"; exit 1; }
     trap "rm -f $connmap" EXIT
     docker exec -it gateway ${installdir}/bin/beerocks_cli -c bml_conn_map > "$connmap"
